pyproject.toml
[tool.poetry]
name = "ghunner"
version = "0.1.0"
description = ""
authors = ["Bjørnar Moe Remmen <bjornarremm@gmail.com>"]

[tool.poetry.scripts]
ghunner = "ghunner.ghunner:ghunner"

[tool.poetry.dependencies]
python = "^3.7"
click = "^8.1.3"
click-default-group = "^1.2.2"
cloup = "^2.0.0"
PyGithub = "^1.58.1"
keyring = "^23.13.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"